+++ tests/Unit/Domain.LinnApps.Tests/ImportBookServiceTests/WhenUpdatingWithInvoiceDetails.cs
@@ -7,50 +7,68 @@
     using Linn.Stores.Domain.LinnApps.ImportBooks;
     using NUnit.Framework;
 
-
     public class WhenUpdatingWithInvoiceDetails : ContextBase
     {
         private readonly int impbookId = 12007;
-
         private ImportBook impbook;
 
         [SetUp]
         public void SetUp()
         {
-            var firstInvoiceDetail = new ImportBookInvoiceDetail() { ImportBookId = this.impbookId, InvoiceNumber = "123", LineNumber = 1, InvoiceValue = (decimal)12.5 };
+            var firstInvoiceDetail = new ImportBookInvoiceDetail
+                                     {
+                                         ImportBookId = this.impbookId,
+                                         InvoiceNumber = "123",
+                                         LineNumber = 1,
+                                         InvoiceValue = 12.5m
+                                     };
 
-            var secondInvoiceDetail = new ImportBookInvoiceDetail() { ImportBookId = this.impbookId, InvoiceNumber = "1234", LineNumber = 2, InvoiceValue = (decimal)155.2 };
+            var secondInvoiceDetail = new ImportBookInvoiceDetail
+                                      {
+                                          ImportBookId = this.impbookId,
+                                          InvoiceNumber = "1234",
+                                          LineNumber = 2,
+                                          InvoiceValue = 155.2m
+                                      };
 
-            var updatedFirstInvoiceDetail = new ImportBookInvoiceDetail() { ImportBookId = this.impbookId, InvoiceNumber = "133", LineNumber = 1, InvoiceValue = (decimal)125.5 };
-
+            var updatedFirstInvoiceDetail = new ImportBookInvoiceDetail
+                                            {
+                                                ImportBookId = this.impbookId,
+                                                InvoiceNumber = "133",
+                                                LineNumber = 1,
+                                                InvoiceValue = 125.5m
+                                            };
 
             this.impbook = new ImportBook
-            {
-                Id = this.impbookId,
-                DateCreated = DateTime.Now.AddDays(-5),
-                SupplierId = 555,
-                CarrierId = 678,
-                TransportId = 1,
-                TransactionId = 44,
-                TotalImportValue = (decimal)123.4,
-                InvoiceDetails = new List<ImportBookInvoiceDetail> { firstInvoiceDetail },
-                OrderDetails = new List<ImportBookOrderDetail>(),
-                PostEntries = new List<ImportBookPostEntry>()
-            };
+                           {
+                               Id = this.impbookId,
+                               DateCreated = DateTime.Now.AddDays(-5),
+                               SupplierId = 555,
+                               CarrierId = 678,
+                               TransportId = 1,
+                               TransactionId = 44,
+                               TotalImportValue = 123.4m,
+                               InvoiceDetails = new List<ImportBookInvoiceDetail> { firstInvoiceDetail },
+                               OrderDetails = new List<ImportBookOrderDetail>(),
+                               PostEntries = new List<ImportBookPostEntry>()
+                           };
 
             var newImportBook = new ImportBook
-            {
-                Id = this.impbookId,
-                DateCreated = DateTime.Now.AddDays(-5),
-                SupplierId = 555,
-                CarrierId = 678,
-                TransactionId = 44,
-                TotalImportValue = (decimal)123.4,
-                InvoiceDetails = new List<ImportBookInvoiceDetail> { updatedFirstInvoiceDetail, secondInvoiceDetail },
-                OrderDetails = new List<ImportBookOrderDetail>(),
-                PostEntries = new List<ImportBookPostEntry>()
-            };
-
+                                {
+                                    Id = this.impbookId,
+                                    DateCreated = DateTime.Now.AddDays(-5),
+                                    SupplierId = 555,
+                                    CarrierId = 678,
+                                    TransactionId = 44,
+                                    TotalImportValue = 123.4m,
+                                    InvoiceDetails =
+                                        new List<ImportBookInvoiceDetail>
+                                        {
+                                            updatedFirstInvoiceDetail, secondInvoiceDetail
+                                        },
+                                    OrderDetails = new List<ImportBookOrderDetail>(),
+                                    PostEntries = new List<ImportBookPostEntry>()
+                                };
 
             this.Sut.Update(this.impbook, newImportBook);
         }
@@ -59,8 +77,8 @@
         public void ShouldHaveUpdatedInvoiceDetail()
         {
             this.impbook.InvoiceDetails.FirstOrDefault(
-                    x => x.LineNumber == 1 && x.InvoiceNumber == "133" && x.InvoiceValue == (decimal)125.5)
-                .Should().NotBeNull();
+                    x => x.LineNumber == 1 && x.InvoiceNumber == "133" && x.InvoiceValue == 125.5m).Should()
+                .NotBeNull();
         }
 
         [Test]
@@ -68,8 +86,8 @@
         {
             this.impbook.InvoiceDetails.Count().Should().Be(2);
             this.impbook.InvoiceDetails.FirstOrDefault(
-                    x => x.LineNumber == 2 && x.InvoiceNumber == "1234" && x.InvoiceValue == (decimal)155.2)
-                .Should().NotBeNull();
+                    x => x.LineNumber == 2 && x.InvoiceNumber == "1234" && x.InvoiceValue == 155.2m).Should()
+                .NotBeNull();
         }
     }
 }
